+++ bootstrap.ml
@@ -41,6 +41,7 @@ let dirs =
   ; "src/ocaml-config"              , Some "Ocaml_config"
   ; "vendor/boot"                   , None
   ; "src/dune_lang"                 , Some "Dune_lang"
+  ; "otherlibs/build-info/src"      , None
   ; "src"                           , None
   ]