+++ otherlibs/build-info/src/build_info.mli
@@ -1,0 +1,29 @@
+(** Provide build information *)
+
+module V1 : sig
+
+  module Version : sig
+    type t
+
+    val to_string : t -> string
+  end
+
+  (** The version at which the current executable was built. The
+      version is [None] during development, it is only [Some _] once
+      the executable is installed or promoted to the source tree. *)
+  val version : Version.t option
+
+  module Statically_linked_library : sig
+    type t
+
+    (** The most visible name of the library. If it is a public
+        library, this its public name otherwise it is its private name.
+    *)
+    val name : t -> string
+
+    val version : t -> Version.t option
+  end
+
+  (** All the libraries that where statically linked in *)
+  val statically_linked_libraries : Statically_linked_library.t list
+end